Adopted typical Ruby conventions for test running.
This commit restructures the test case files and
removes "test-api.rb" in favor of running tests
via Rake, with helper and initialization code
defined in the "test_helper.rb" file.

It also…

* renames all test case filenames to end with
  "_test.rb", to aid in filesystem globbing;

* no longer requires you to include new test cases
  into a parent TestCase instance. Simply create a
  MiniTest::Unit::TestCase instance for each test
  case.
