+++ .gitignore
@@ -1,4 +1,5 @@
 *.gem
+*.lock
 *.rbc
 .bundle
 .config
