+++ orchestrate-api.gemspec
@@ -18,6 +18,8 @@ Gem::Specification.new do |s|
 	s.test_files    = s.files.grep(%r{^(test|spec|features)/})
 	s.require_paths = ["lib"]
 
+  s.add_development_dependency "bundler", "~> 1.6"
+  s.add_development_dependency "rake"
 	s.add_development_dependency "vcr"
 	s.add_development_dependency "webmock"
